deps: nuxt rc9 & desmjs 4.3.5
