+++ package.json
@@ -13,9 +13,9 @@
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^3.14.1",
-    "@desmoslabs/desmjs": "4.3.4",
-    "@desmoslabs/desmjs-types": "^4.3.4",
-    "@desmoslabs/desmjs-walletconnect": "^4.3.4",
+    "@desmoslabs/desmjs": "^4.3.5",
+    "@desmoslabs/desmjs-types": "^4.3.5",
+    "@desmoslabs/desmjs-walletconnect": "^4.3.5",
     "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
     "@ezone-devops/milkdown-plugin-directive-fallback": "^6.2.2",
     "@keplr-wallet/types": "0.10.19",
@@ -41,13 +41,13 @@
     "@typescript-eslint/eslint-plugin": "^5.30.6",
     "@typescript-eslint/parser": "^5.30.6",
     "c8": "^7.11.3",
-    "cosmjs-types": "^0.5.0",
+    "cosmjs-types": "^0.5.1",
     "eslint": "^8.20.0",
     "eslint-plugin-vue": "^9.2.0",
     "ipfs-http-client": "^57.0.3",
     "katex": "^0.16.0",
     "miniflare": "^2.6.0",
-    "nuxt": "^3.0.0-rc.6",
+    "nuxt": "npm:nuxt3@3.0.0-rc.9-27699566.0339a23",
     "playwright": "^1.23.4",
     "prosemirror-model": "^1.18.1",
     "remark-directive": "^2.0.1",
@@ -58,15 +58,18 @@
   },
   "dependencies": {
     "@nuxtjs/composition-api": "^0.33.0",
-    "@pinia/nuxt": "^0.3.0",
+    "@pinia/nuxt": "^0.4.1",
     "@rollup/plugin-inject": "^4.0.4",
+    "browser": "^0.2.6",
     "buffer": "^6.0.3",
     "bufferutil": "^4.0.6",
-    "pinia": "^2.0.16",
+    "events": "^3.3.0",
+    "pinia": "^2.0.21",
     "qrcode.vue": "^3.3.3",
     "stream-browserify": "^3.0.0",
     "unique-username-generator": "^1.1.1",
     "utf-8-validate": "^5.0.9",
+    "util": "^0.12.4",
     "uuid": "^8.3.2"
   }
 }
